CamStack_1.0.0/camstack/pyproject.toml
[project]
name = "camstack"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "jinja2",
  "python-multipart",
  "loguru",
  "wsdiscovery",
  "onvif-zeep",
  "psutil",
  "requests",
  "starlette",
]

[tool.uv]
index-url = "https://pypi.org/simple"
